feat(Clients): Directory checks

+++ Copy/Clients/FTP.cs
@@ -50,6 +50,12 @@ namespace Copy.Clients
 
         public Stream GetFile(string path)
         {
+            string directory = Path.GetDirectoryName(path) ?? throw new ArgumentNullException($"Impossible to get directory from {path}");
+            if (!FtpClient.DirectoryExists(directory))
+            {
+                Logger.Error($"Directory {directory} does not exist");
+                throw new DirectoryNotFoundException($"Directory {directory} does not exist");
+            }
             if (!DoFileExist(path))
             {
                 Logger.Error($"File {path} does not exist");
@@ -64,6 +70,12 @@ namespace Copy.Clients
 
         public void PutFile(string path, Stream stream)
         {
+            string directory = Path.GetDirectoryName(path) ?? throw new ArgumentNullException($"Impossible to get directory from {path}");
+            if (!FtpClient.DirectoryExists(directory))
+            {
+                Logger.Warn($"Directory {directory} does not exist, creating");
+                FtpClient.CreateDirectory(directory);
+            }
             if (DoFileExist(path)) Logger.Warn($"File {path} already exists, overwriting");
 
             FtpClient.UploadStream(stream, path, FtpRemoteExists.Overwrite);
@@ -71,6 +83,12 @@ namespace Copy.Clients
 
         public void DeleteFile(string path)
         {
+            string directory = Path.GetDirectoryName(path) ?? throw new ArgumentNullException($"Impossible to get directory from {path}");
+            if (!FtpClient.DirectoryExists(directory))
+            {
+                Logger.Error($"Directory {directory} does not exist");
+                throw new DirectoryNotFoundException($"Directory {directory} does not exist");
+            }
             if (!DoFileExist(path)) Logger.Warn($"File {path} does not exist");
 
             FtpClient.DeleteFile(path);
